♻️ refactor: Update ProductAttributes and DTOs to use float64 for ExpirationRate and FreezingRate, and fix SQL query in productIdExists method

+++ internal/domain/dto/product_dto.go
@@ -13,9 +13,9 @@ type ProductRequestDTO struct {
 	Height                         float64 `json:"height"`
 	Length                         float64 `json:"length"`
 	NetWeight                      float64 `json:"net_weight"`
-	ExpirationRate                 int     `json:"expiration_rate"`
+	ExpirationRate                 float64 `json:"expiration_rate"`
 	RecommendedFreezingTemperature float64 `json:"recommended_freezing_temperature"`
-	FreezingRate                   int     `json:"freezing_rate"`
+	FreezingRate                   float64 `json:"freezing_rate"`
 	ProductTypeId                  int     `json:"product_type_id"`
 	SellerId                       int     `json:"seller_id"`
 }
@@ -28,9 +28,9 @@ type ProductResponseDTO struct {
 	Height                         float64 `json:"height"`
 	Length                         float64 `json:"length"`
 	NetWeight                      float64 `json:"net_weight"`
-	ExpirationRate                 int     `json:"expiration_rate"`
+	ExpirationRate                 float64 `json:"expiration_rate"`
 	RecommendedFreezingTemperature float64 `json:"recommended_freezing_temperature"`
-	FreezingRate                   int     `json:"freezing_rate"`
+	FreezingRate                   float64 `json:"freezing_rate"`
 	ProductTypeId                  int     `json:"product_type_id"`
 	SellerId                       int     `json:"seller_id"`
 }
@@ -42,9 +42,9 @@ type ProductRequestDTOPtr struct {
 	Height                         *float64 `json:"height"`
 	Length                         *float64 `json:"length"`
 	NetWeight                      *float64 `json:"net_weight"`
-	ExpirationRate                 *int     `json:"expiration_rate"`
+	ExpirationRate                 *float64 `json:"expiration_rate"`
 	RecommendedFreezingTemperature *float64 `json:"recommended_freezing_temperature"`
-	FreezingRate                   *int     `json:"freezing_rate"`
+	FreezingRate                   *float64 `json:"freezing_rate"`
 	ProductTypeId                  *int     `json:"product_type_id"`
 	SellerId                       *int     `json:"seller_id"`
 }
